+++ crates/bevy_ui/src/ui_node.rs
@@ -31,10 +31,10 @@ impl Node {
 
     /// Returns the size of the node in physical pixels based on the given scale factor.
     #[inline]
-    pub fn physical_size(&self, scale_factor: f64) -> Vec2 {
+    pub fn physical_size(&self, scale_factor: f32) -> Vec2 {
         Vec2::new(
-            (self.calculated_size.x as f64 * scale_factor) as f32,
-            (self.calculated_size.y as f64 * scale_factor) as f32,
+            self.calculated_size.x * scale_factor,
+            self.calculated_size.y * scale_factor,
         )
     }
 
@@ -46,17 +46,11 @@ impl Node {
 
     /// Returns the physical pixel coordinates of the UI node, based on its [`GlobalTransform`] and the scale factor.
     #[inline]
-    pub fn physical_rect(&self, transform: &GlobalTransform, scale_factor: f64) -> Rect {
+    pub fn physical_rect(&self, transform: &GlobalTransform, scale_factor: f32) -> Rect {
         let rect = self.logical_rect(transform);
         Rect {
-            min: Vec2::new(
-                (rect.min.x as f64 * scale_factor) as f32,
-                (rect.min.y as f64 * scale_factor) as f32,
-            ),
-            max: Vec2::new(
-                (rect.max.x as f64 * scale_factor) as f32,
-                (rect.max.y as f64 * scale_factor) as f32,
-            ),
+            min: Vec2::new(rect.min.x * scale_factor, rect.min.y * scale_factor),
+            max: Vec2::new(rect.max.x * scale_factor, rect.max.y * scale_factor),
         }
     }
 }
